% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dconcordant.R
\name{dconcordant}
\alias{dconcordant}
\title{Calculate the probability atom of the count of concordant pairs among indpendent pairs of random variables.}
\usage{
dconcordant(ss, nn)
}
\arguments{
\item{ss}{Integer number of pairs}

\item{nn}{number of pairs}
}
\value{
real probability
}
\description{
Calculate the probability atom of the count of concordant pairs among indpendent pairs of random variables.
}
