\name{dmannwhitney}
\alias{dmannwhitney}

\title{Probabilities associated with the Mann-Whitney test}
\description{The function estimates the probability associated with Mann-Whitney test.}

\usage{dmannwhitney(u, m, n)}

\arguments{
  \item{u}{ Mann-Whitney statistic}
  \item{m}{ number of observation in the first group}
  \item{n}{ number of observation in the second group}
}


%\value{}

\references{J.J. Higgins, (2004), \emph{Introduction to Modern Nonparametric Statistics}, Brooks/Cole, Cengage Learning.}

\author{John E. Kolassa}
