% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecdfcis.R
\name{ecdfcis}
\alias{ecdfcis}
\title{Confidence Intervals for Empirical Cumulative Distribution Functions}
\usage{
ecdfcis(data, alpha = 0.05, dataname = NA, exact = TRUE, newplot = TRUE)
}
\arguments{
\item{data}{vector of observations}

\item{alpha}{1-confidence level.}

\item{dataname}{Name of variable for use in axis labeling}

\item{exact}{logical value controlling whether confidence intervals are exact or asymlptotic.}

\item{newplot}{logical value controlling whether the estimate is added to an existing plot, or whether a new plot should be constructed.}
}
\description{
Confidence Intervals for Empirical Cumulative Distribution Functions
}
