% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exactquantileci.R
\name{exactquantileci}
\alias{exactquantileci}
\title{Exact Quantile Confidence Interval}
\usage{
exactquantileci(xvec, tau = 0.5, alpha = 0.05, md = 0)
}
\arguments{
\item{xvec}{vector of observations}

\item{tau}{quantile to be estimated.  If this is a vector, separate intervals and tests for each value will be calculated.}

\item{alpha}{1-confidence level.}

\item{md}{null value of quantile}
}
\value{
A list with components cis, an array with two columns, representing lower and upper bounds, and a vector pvals, of p-values.
}
\description{
Calculates exact quanitle confidence intervals by inverting the generalization of the sign test.
}
