% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genscorestat.R
\name{genscorestat}
\alias{genscorestat}
\title{Normal-theory two sample scorestatistic.}
\usage{
genscorestat(scores, group, correct = 0)
}
\arguments{
\item{scores}{scores of the data.}

\item{group}{numeric or character vector of group identities.}

\item{correct}{half the minimal distance between two potential values of the score statistic.}
}
\value{
Object of class htest containing the p-value.
}
\description{
Calculates the p-value from the normal approximation to the permutation distribution of a two-sample score statistic.
}
