% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mannwhitney.test.R
\name{mannwhitney.test}
\alias{mannwhitney.test}
\title{Perform the Mann Whitney two-sample test}
\usage{
mannwhitney.test(x, y, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{x}{A vector of values from the first sample.}

\item{y}{A vector of values from the first sample.}

\item{alternative}{Specification of alternative hypothesis.}
}
\value{
Test results of class htest
}
\description{
Perform the Mann Whitney two-sample test
}
\examples{
mannwhitney.test(rnorm(10),rnorm(10)+.5)
}
