% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultNonParam.R
\docType{package}
\name{MultNonParam-package}
\alias{MultNonParam}
\alias{MultNonParam-package}
\title{MultNonParam}
\description{
A collection of nonparametric methods.
}
\author{
\strong{Maintainer}: John E. Kolassa \email{kolassa@stat.rutgers.edu}

Authors:
\itemize{
  \item Stephane Jankowski
}

}
