% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nextp.R
\name{nextp}
\alias{nextp}
\title{Next permutation}
\usage{
nextp(perm, b = 1)
}
\arguments{
\item{perm}{indices to be permutedj}

\item{b}{number to begin at.  Set equal to 1.}
}
\value{
The next permutation
}
\description{
Cycles through permutations of first argument
}
