% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/page.test.unbalanced.R
\name{page.test.unbalanced}
\alias{page.test.unbalanced}
\title{Perform Page test for unbalanced two-way design}
\usage{
page.test.unbalanced(x, trt, blk, sides = 2)
}
\arguments{
\item{x}{A vector of responses}

\item{trt}{A vector of consecutive integers starting at 1 indicating treatment}

\item{blk}{A vector of consecutive integers starting at 1 indicating block}

\item{sides}{A single integer indicating sides.  Defaults to 2.}
}
\value{
P-value for Page test.
}
\description{
Perform Page test for unbalanced two-way design
}
\examples{
page.test.unbalanced(rnorm(15),rep(1:3,5),rep(1:5,rep(3,5)))
}
