% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pconcordant.R
\name{pconcordant}
\alias{pconcordant}
\title{Calculate the cumulative distribution of the count of concordant pairs among indpendent pairs of random variables.}
\usage{
pconcordant(ss, nn)
}
\arguments{
\item{ss}{Integer number of pairs}

\item{nn}{number of pairs}
}
\value{
real probability
}
\description{
Calculate the cumulative distribution of the count of concordant pairs among indpendent pairs of random variables.
}
