% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/powerplot.R
\name{powerplot}
\alias{powerplot}
\title{Power Plot}
\usage{
powerplot(
  numgrps = 3,
  thetadagger = NULL,
  nnvec = 5:30,
  nmc = 50000,
  targetpower = 0.8,
  level = 0.05
)
}
\arguments{
\item{numgrps}{Number of groups to compare}

\item{thetadagger}{Direction of effect}

\item{nnvec}{vector of numbers per group.}

\item{nmc}{Number of Monte Carlo trials}

\item{targetpower}{Target power for test}

\item{level}{level for test.}
}
\description{
Plots powers for the Kruskall-Wallis test, via Monte Carlo and two approximations.
}
