% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwiseprobabilities.R
\name{probabilityderiv}
\alias{probabilityderiv}
\title{Derivative of pairwise probabilities of Exceedence}
\usage{
probabilityderiv(distname = c("normal", "logistic"))
}
\arguments{
\item{distname}{The distribution of the underlying observations; normal and logistic are currently supported.}
}
\value{
The scalar derivative.
}
\description{
\code{probabilityderiv} calculates derivatives probabilities of one variable exceeding another,
where the variables are independent, and with identical distributions  except for a location shift, at the null hypothesis.
This calculation is useful for power of Mann-Whitney-Wilcoxon, Jonckheere-Terpstra, and Kruskal-Wallis testing.
}
\details{
Probabilities of particular families must be calculated analytically, and then differentiated.
}
