% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prostate.R
\docType{data}
\name{prostate}
\alias{prostate}
\title{prostate}
\format{\itemize{
\item hosp : Hospital in which the patient is hospitalized.
\item stage : stage of the cancer.
\item gleason score : used to help evaluate the prognosis of the cancer.
\item psa : prostate-specific antigen.
\item age : age of the patient.
\item advanced : boolean. \code{TRUE} if the cancer is advanced.
}}
\usage{
data(prostate)
}
\description{
221 prostate cancer patients are collected in this data set.
}
\examples{
data(prostate)
attach(prostate)
plot(age,psa,main="Age and PSA",sub="Prostate Cancer Data",
   xlab="Age (years)",ylab="PSA")
}
\references{
A. V. D'Amico, R. Whittington, S. B. Malkowicz, D. Schultz, K. Blank, G. A. Broderick, J. E. Tomaszewski, A. A. Renshaw, I. Kaplan, C. J. Beard, A. Wein (1998) , \emph{Biochemical outcome after radical prostatectomy, external beam radiation therapy, or interstitial radiation therapy for clinically localized prostate cancer}, JAMA : the journal of the American Medical Association 280 969-74.
}
