% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qconcordant.R
\name{qconcordant}
\alias{qconcordant}
\title{Calculate the quantiles of the count of concordant pairs among indpendent pairs of random variables.}
\usage{
qconcordant(qq, nn, exact = TRUE)
}
\arguments{
\item{qq}{Desired quantile}

\item{nn}{number of pairs}

\item{exact}{flag to trigger exact calculation when possible.}
}
\value{
Integer quantile
}
\description{
Calculate the quantiles of the count of concordant pairs among indpendent pairs of random variables.
}
