% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sensativity.plot.R
\name{sensitivity.plot}
\alias{sensitivity.plot}
\title{Compare the sensitivity of different statistics.}
\usage{
sensitivity.plot(y, sub, stats)
}
\arguments{
\item{y}{vector of the data.}

\item{sub}{subtitle for the plot.}

\item{stats}{vector of functions to be plotted.}
}
\description{
Compare the sensitivity of different statistics.
}
\details{
To compare the sensitivity, outliers are added to the original data. The shift of each statistics due to the new value is measured and plotted.
}
