% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftcr.R
\name{shiftcr}
\alias{shiftcr}
\title{Nonparametric Confidence Region for a Vector Shift Parameter}
\usage{
shiftcr(xm,  hpts = 50)
}
\arguments{
\item{xm}{A two-column matrix of bivariate data whose two location parameters are to be estimated.}
\item{hpts}{Controls the number of grid points, by constructing a grid of 2*hpts+1 on each side.}
}
\value{
nothing
}
\description{
Inversion of a one-sample bivariate rank test is used to produce a confidence region.  The region is constructed by building a grid of potential parameter values, evaluating the test statistic on each grid point, collecting the p-values, and then drawing the appropriate countour of the p-values.  The grid is centered at the bivariate median of the data set.
}
