% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solvencp.R
\name{solvencp}
\alias{solvencp}
\title{Noncentrality Parameter for a Given Level and Power}
\usage{
solvencp(df, level = 0.05, targetpower = 0.8)
}
\arguments{
\item{df}{Common degrees of freedom for null and alternative distributions.}

\item{level}{Level (that is, type I error rate) for the test.}

\item{targetpower}{Desired power}
}
\value{
required noncentrality parameter.
}
\description{
This function calculates the noncentrality parameter required to give a test whose null distribution is central chi-square and whose alternative distribution is noncentral chi-square the required level and power.
}
\examples{
solvencp(4)
}
