% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terpstra.test.R
\name{terpstra.test}
\alias{terpstra.test}
\title{Perform the Terpstra version of the multi-ordered-sample test}
\usage{
terpstra.test(x, g, alternative = c("two.sided", "less", "greater"))
}
\arguments{
\item{x}{A vector of values from all samples.}

\item{g}{A vector of group labels.}

\item{alternative}{Specification of alternative hypothesis.}
}
\value{
Test results of class htest
}
\description{
Perform the Terpstra version of the multi-ordered-sample test
}
\examples{
terpstra.test(rnorm(15),rep(1:3,5))
}
