% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theil.R
\name{theil}
\alias{theil}
\title{Perform the Theil nonparametric estimation and confidence interval for a slope parameter.}
\usage{
theil(x, y, conf = 0.9)
}
\arguments{
\item{x}{A vector of values of the explanatory variable.}

\item{y}{A vector of values of the response variable.}

\item{conf}{Level of confidence interval.}
}
\value{
A list with letters and numbers.
\itemize{
  \item est - An estimate, the median of pairwise slopes.
  \item ci - A vector of confidence interval endpoints.
}
}
\description{
Perform the Theil nonparametric estimation and confidence interval for a slope parameter.
}
\examples{
a<-0:19;b<-a^2.5
theil(a,b)
}
