% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tukey.kruskal.test.R
\name{tukey.kruskal.test}
\alias{tukey.kruskal.test}
\title{Tukey HSD procedure}
\usage{
tukey.kruskal.test(resp, grp, alpha = 0.05)
}
\arguments{
\item{resp}{vector containing the values for the variable of interest.}

\item{grp}{vector specifying in which group is each observation.}

\item{alpha}{level of the test.}
}
\value{
A logical vector for every combinaison of two groups. \code{TRUE} if the distribution in one group is significantly different from the distribution in the other group.
}
\description{
Rank-based method for controlling experiment-wise error. Assume normality of the distribution for the variable of interest.
}
\details{
The original Tuckey HSD procedure is supposed to be applied for equal sample sizes. However, the \code{tukey.kruskal.test} function performs the Tukey-Kramer procedure that works for unequal sample sizes.
}
\references{
J.J. Higgins, (2004), \emph{Introduction to Modern Nonparametric Statistics}, Brooks/Cole, Cengage Learning.
}
