% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/twosamplesurvtests.R
\name{twosamplesurvtests}
\alias{twosamplesurvtests}
\title{Two Sample Omnibus Tests of Survival Curves}
\usage{
twosamplesurvtests(times, delta, grp)
}
\arguments{
\item{times}{Event and censoring times}

\item{delta}{Indicator of event (1) or censoring (0).}

\item{grp}{Variable that divides the population into groups.}
}
\value{
A vector of length two, with the Kolmogorov-Smirnov and Anderson-Darling statistics.
}
\description{
Returns the Kolmogorov-Smirnov and Anderson-Darling test statistics for two right-censored data sets.
}
\examples{
twosamplesurvpvs(rexp(20),rbinom(20,1,.5),rbinom(20,1,.5))
}
