% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.jplot.R
\name{util.jplot}
\alias{util.jplot}
\title{Plot a curve, skipping bits where there is a large jump.}
\usage{
util.jplot(x, y, ...)
}
\arguments{
\item{x}{Ordinates to be plotted.}

\item{y}{Abcissas to be plotted.}

\item{...}{Arguents passed directly to plot.}
}
\description{
Plot a curve, skipping bits where there is a large jump.
}
