% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betatest.R
\name{betatest}
\alias{betatest}
\title{Permutation test of assication}
\usage{
betatest(x, y)
}
\arguments{
\item{x}{First vector to be associated.}

\item{y}{First vector to be associated.}
}
\value{
p-value
}
\description{
Calculate the p-value for the test of association between two variables using the permutation method.
}
\examples{
#Example using data from plant Qn1 from the CO2 data set.^M
betatest(CO2[CO2$Plant=="Qn1",4],CO2[CO2$Plant=="Qn1",5])
}
