% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dmannwhitney.R
\name{dmannwhitney}
\alias{dmannwhitney}
\title{Mann Whitney Probability Mass function}
\usage{
dmannwhitney(u, m, n)
}
\arguments{
\item{u}{Statistic value}

\item{m}{Group 1 size}

\item{n}{Group 2 size}
}
\value{
Probability that the Mann-Whitney statistic takes the value u under H0
}
\description{
Calculates the Mann Whitney Probability Mass function recursively.
}
