% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiftcr.R
\name{shiftcr}
\alias{shiftcr}
\title{Nonparametric Confidence Region for a Vector Shift Parameter}
\usage{
shiftcr(xm)
}
\arguments{
\item{xm}{A bivariate data vector whosse two location parameters are to be estimated.}
}
\value{
nothing
}
\description{
Inversion of a one-sample bivariate rank test is used to produce a confidence region.
}
